cleaned up code + both close autos structure
