+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/BasicAuto.java
@@ -8,10 +8,11 @@ import com.acmerobotics.roadrunner.SequentialAction;
 import com.acmerobotics.roadrunner.SleepAction;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import org.firstinspires.ftc.teamcode.BrainSTEMAutoRobot;
 
-
+@Disabled
 @Autonomous (name = "Auto")
 @Config
 public final class BasicAuto extends LinearOpMode {
