+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto_subsystems/Shooter.java
@@ -7,7 +7,7 @@ import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.Component;
+import org.firstinspires.ftc.teamcode.util.Component;
 
 @Config
 public class Shooter implements Component {
