+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tele_subsystems/Collector.java
@@ -5,7 +5,7 @@ import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.Component;
+import org.firstinspires.ftc.teamcode.util.Component;
 
 
 @Config
