+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tele_subsystems/Finger.java
@@ -7,9 +7,7 @@ import com.qualcomm.robotcore.hardware.ServoImplEx;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.Component;
-
-import java.time.OffsetDateTime;
+import org.firstinspires.ftc.teamcode.util.Component;
 
 
 @Config
