+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tele_subsystems/Spindexer.java
@@ -8,8 +8,8 @@ import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.BrainSTEMTeleRobot;
-import org.firstinspires.ftc.teamcode.Component;
-import org.firstinspires.ftc.teamcode.PIDController;
+import org.firstinspires.ftc.teamcode.util.Component;
+import org.firstinspires.ftc.teamcode.util.PIDController;
 
 
 @Config
