+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/Component.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.util;
 
 public interface Component {
     void reset();
